feat: Add functionality to save additional description for a request
The code changes include adding a new action method `SaveAdditionalDescription` in the `HomeController.cs` file. This method allows saving an additional description for a request and updating its status. The changes also include updating the `RequestDetail.cshtml` file to include a JavaScript function `saveAdditionalDescription` that sends an AJAX request to the `SaveAdditionalDescription` action method.

Based on the recent user commits and repository commits, it seems that this is a new feature being added to the application.