feat: Add "tamirgah" field to user model and update related views
The code changes include adding a new field "tamirgah" to the user model in the `HomeController.cs` file. This field represents the URL of the user's tamirgah. The changes also include updating the `RequestDetails.cshtml` view to display the tamirgah field for the user.

Based on the recent user commits and repository commits, it appears that this commit adds the "tamirgah" field to the user model and updates the related views.